third_party/WebKit/LayoutTests/editing/deleting/delete-across-editable-content-boundaries-3.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
    '<div contenteditable>',
      '<table contenteditable="false"><tbody><tr>',
        '<td contenteditable>foo|bar</td>',
        '<td contenteditable>baz<span contenteditable="false">abc</span>def</td>',
      '</tr></tbody></table>',
      'ghi^jk',
    '</div>'
  ].join(''),
  'delete',
  [
    '<div contenteditable>',
      '<table contenteditable="false"><tbody><tr>',
        '<td contenteditable>foo|jk</td>',
        '<td contenteditable></td>',
      '</tr></tbody></table>',
    '</div>'
  ].join('')),
  'Delete selected editable nodes that are not in both the start and end editable roots');
</script>
